Reduce logic in each instruction class related to addressing modes and cycle count. Use predetermined base cycle count per opcode, and add additional cycles depending on addressing mode and other factors.

+++ Highbyte.DotNet6502/OpCode.cs
@@ -25,6 +25,6 @@
         /// Can take more cycles depending on zero page wrap-around or crossing page boundary
         /// </summary>
         /// <value></value>
-        public int MinimumCycles { get; set;}
+        public ulong MinimumCycles { get; set;}
     }
 }
